+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.7-dev
 
 * Updates test entry points to pure python. README instructions for running tests were updated.
+* Missing configuration files are no longer ignored. A runtime exception is raised immediately when an expected configuration file is not found. 
 
 ## 0.0.6